% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapnet-package.R
\docType{package}
\name{tapnet-package}
\alias{tapnet-package}
\title{tapnet: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

