% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapnet-package.R
\docType{package}
\name{tapnet-package}
\alias{tapnet-package}
\title{tapnet: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
making tapnet objects, fitting the tapnet, and predicting to new data. All other functions are helpers or quality monitoring.
}
\section{Making a tapnet object}{

To be filled with content ... For now, see example in predict_tapnet.
Alternatively to having actual observed data, one can simulate a tapnet object.
}

